move projects fetch to server
