+++ src/components/projects/project.tsx
@@ -1,0 +1,10 @@
+import { Repo } from '@/lib/data/repositories';
+
+export function Project({ repo }: { repo: Repo }) {
+  return (
+    <div>
+      <p>{repo.name}</p>
+      <p>{repo.description}</p>
+    </div>
+  );
+}
