--- src/lib/queries/index.ts
@@ -1,1 +1,0 @@
-export * from './repositories';
